+++ src/main/java/br/com/totvs/api/veiculo/ficticiusClean/service/CalculoConsumoService.java
@@ -1,6 +1,6 @@
 package br.com.totvs.api.veiculo.ficticiusClean.service;
 
 public interface CalculoConsumoService {
-    float calcularPrecoTotal (float precoCombustivel, float consumoCombustivel);
-    float calcularUso (float distanciaPercorrida, float consumoMedia);
+    Double calcularPrecoTotal (Double precoCombustivel, Double consumoCombustivel);
+    Double calcularUso (Double distanciaPercorrida, Double consumoMedia);
 }
